requirements.txt
Flask
Requests
Pillow
python-dotenv
